// src/modules-system/prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const datphong_trang_thai = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled'
} as const

export type datphong_trang_thai = (typeof datphong_trang_thai)[keyof typeof datphong_trang_thai]